Promote the escalation path resource out of beta
incident_escalation_path is now the resource that was
incident_escalation_path_beta, and the one that nested a branch's nodes
under if_else.then_path and if_else.else_path is removed, along with its
data source.

Nesting was the problem the beta shape existed to solve: Terraform
schemas cannot recurse indefinitely, so a path stopped at five levels of
branching and one that branched deeper could not be written at all.
Sequences sit at one depth and name each other, so there is no limit.

The resource takes a moved block from the name it went by. Coming off
the old shape is a rewrite - branches become sequences, repeat becomes
loop, raw engine conditions become branch.if - but the escalation path
itself is untouched, so it imports by ID and everything escalating
through it carries on.

Deleting the old resource meant finding what else was living in its
files. levelAttrTypes and friends were already shared; the node models,
the target helpers, decodeTargets, the schedule-mode constant and the
validate warnings were not, and now sit in escalation_path_nodes.go and
the promoted resource beside the code that uses them. The data source
was looking the Escalation Path catalog type up by borrowing the old
data source's method, and now does it itself.

Three tests went with the shape they tested: the nested round-trip, the
node type/block agreement check (the new shape has no type attribute),
and the unknown-condition-binding regression from ONC (a branch's
condition has no param bindings to leave unknown). The check that both
resources built their shared node blocks from one definition goes too -
there is only one resource to check now. teamTypeName, which several
other acceptance tests use, moved to provider_test.go rather than
disappearing with the file it happened to live in.
