+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- **Breaking**: `incident_escalation_path_beta` is now `incident_escalation_path`, and the resource that nested a branch's nodes under `if_else.then_path` and `if_else.else_path` is removed, along with its data source. An escalation path is written as a flat map of named sequences from here on: a `branch` names the sequences to continue down instead of holding their nodes, `loop` replaces `repeat`, and `branch.if` replaces the raw engine `conditions` with one attribute per thing an escalation can be tested on. The five-level limit on branching goes with the nesting, so a path too deep to write before can be written now. The promoted resource takes a `moved` block from `incident_escalation_path_beta`, so anyone already on it renames and applies a plan that reports moves and nothing else. Coming off the old shape is a rewrite, but the escalation path itself is untouched: import it by ID rather than recreating it and everything escalating through it carries on, as the [migration guide](https://registry.terraform.io/providers/incident-io/incident/latest/docs/guides/migrating-to-v7) sets out.
 - **Breaking**: `incident_schedule_beta` and `incident_schedule_rotation_beta` are now `incident_schedule` and `incident_schedule_rotation`, and the `incident_schedule` that declared its rotations inline is removed, along with its data source. A schedule and its rotations are separate resources from here on, so a change to one rotation no longer rewrites the whole schedule - which is what used to disturb who was on call elsewhere on it, or drop overrides attached to a shift. Both promoted resources take a `moved` block from the name they went by, so anyone already on the beta resources renames them and applies a plan that reports moves and nothing else. Coming off the old schema is not a rename, because one resource becomes a schedule plus a rotation apiece: do that on v6, where both schemas are registered, then upgrade - the [migration guide](https://registry.terraform.io/providers/incident-io/incident/latest/docs/guides/migrating-to-v7) has the field mapping and the import IDs. `handover_start_at` is now the rotation's `first_interval_starts_at`, and a rotation's `layers` are gone: a rotation with two layers is one rotation with `concurrent_shifts = 2`. `users = []` is rejected where the old schema accepted it - a rotation with nobody in it is `users = ["NOBODY"]`. `incident_schedule_replica`, `incident_schedule_sync_rule` and `incident_schedule_sync_target` are unchanged, though a replica source still names a layer, which the new rotation resource has no way to give you: replicating a Terraform-managed schedule needs a layer ID from elsewhere until that is addressed.
 - Add a [Migrating to v7](https://registry.terraform.io/providers/incident-io/incident/latest/docs/guides/migrating-to-v7) guide, which is what v7 needs to be a version people can get to. It covers both moves: from the beta resources to the names they keep in v7, which is a rename and a `moved` block per resource, and from the v6 schemas to the ones that replace them, which is a rewrite plus an import because one `incident_schedule` becomes a schedule and a rotation apiece, and one `incident_alert_source` becomes a source and an attribute apiece. The schema work is written to be done on v6, where both schemas exist side by side: upgrading first leaves state under a reused resource name that v7 cannot read, and the guide says how to recover from that too. It also spells out that the old resource must be forgotten with `destroy = false` rather than destroyed, since both resources point at the same object and destroying one deletes the schedule and everyone's shifts.
 - `incident_api_key` and `incident_secret` now claim the key or secret they manage, the way every other managed resource does. Neither did before, so a key or secret Terraform created still looked dashboard-managed: the dashboard offered to edit it, and the next apply put its own configuration back over the change. Both now send the `incident.io/terraform/version` annotation, so the dashboard shows the resource as synced, refuses to edit it, and offers to disconnect one you want to hand back. The claim happens on create, on update, and on import - the last subject to `mark_imported_resources_as_managed`, like every other resource. It needs a version of incident.io that knows `api_key` and `secret` as managed-resource types; against an older one the claim is rejected and the apply reports it.
