+++ examples/resources/incident_escalation_path/import-by-string-id.tf
@@ -1,6 +1,6 @@
 # Import an escalation path using its ID
 # Replace the ID with a real ID from your incident.io organization
 import {
-  to = incident_escalation_path.example
+  to = incident_escalation_path.urgent_support
   id = "01ABC123DEF456GHI789JKL"
 }
