+++ examples/resources/incident_escalation_path/import.sh
@@ -2,4 +2,4 @@
 
 # Import an escalation path using its ID
 # Replace the ID with a real ID from your incident.io organization
-terraform import incident_escalation_path.example 01ABC123DEF456GHI789JKL
+terraform import incident_escalation_path.urgent_support 01ABC123DEF456GHI789JKL
