--- examples/resources/incident_escalation_path_beta/import.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-# Import an escalation path using its ID
-# Replace the ID with a real ID from your incident.io organization
-terraform import incident_escalation_path_beta.urgent_support 01ABC123DEF456GHI789JKL
