+++ internal/provider/provider.go
@@ -162,8 +162,7 @@ func (p *IncidentProvider) Resources(ctx context.Context) []func() resource.Reso
 		NewIncidentCatalogTypeAttributesResource,
 		NewIncidentCustomFieldResource,
 		NewIncidentCustomFieldOptionResource,
-		NewIncidentEscalationPathResource,
-		NewEscalationPathBetaResource,
+		NewEscalationPathResource,
 		NewIncidentRoleResource,
 		NewIncidentIncidentTemplateResource,
 		NewIncidentMaintenanceWindowResource,
@@ -193,8 +192,7 @@ func (p *IncidentProvider) DataSources(ctx context.Context) []func() datasource.
 		NewIncidentCatalogTypeAttributeDataSource,
 		NewIncidentCustomFieldDataSource,
 		NewIncidentCustomFieldOptionDataSource,
-		NewIncidentEscalationPathDataSource,
-		NewEscalationPathBetaDataSource,
+		NewEscalationPathDataSource,
 		NewIncidentRoleDataSource,
 		NewIncidentIncidentTemplateDataSource,
 		NewIncidentTimestampDataSource,
